# openapi.yaml
...
openapi: 3.0.3
info:
  title: Medium Public Unofficial API
  description: >-
    
    This is an unofficial Medium API used to fetch public data such as

    users, articles, lists, etc...





    This API is heavily inspired by [Nishu Jain's Medium API](https://mediumapi.com/), in fact it's nearly the same, I just wanted to make it open source and free.





    It is done using Medium's "official" API - https://medium.com/_/api (I quote official because there is no documentation for this API and no way of knowing it exists without using Developer Tools in the web browser) - and some web scraping.





    Useful links:



    - [API's repository](https://github.com/estebanthi/medium-api)



    - [OpenAPI specification](https://github.com/estebanthi/medium-api/blob/master/openapi.yaml)
  contact:
    email: contact@esteban-thilliez.com
  version: "1.0"
servers:
  - url: https://mediumapi.esteban-thilliez.com/
...
paths:
...
  "/publications/id-for/{publicationSlug}":
    get:
      summary: Get Publication ID
      tags:
        - Publications
      description: >
        
        Returns the unique `publicationId` for the given `publicationSlug`.





        **Note:** You can find the `publication_slug` from the publication's homepage URL.



        - medium.com/**publicationSlug**
      parameters:
        - name: publicationSlug
          in: path
          required: true
          description: It's a lowercased hyphen-separated unique string alloted to each
            Medium Publication.
          schema:
            type: string
            example: illumination
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: string
                    description: Unique hash id of the publication.
                    example: eca1ba5ae1ca
  "/publications/{publicationId}":
    get:
      summary: Get Publication Info
      tags:
        - Publications
      description: >-
        
        Returns publication related information like *Name, Description, Followers

        count, publication ID, etc...*





        It takes `publicationId` as the path parameter.
      parameters:
        - name: publicationId
          in: path
          required: true
          description: It's a unique hash id assigned to every single Medium publication.
          schema:
            type: string
            example: eca1ba5ae1ca
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      publicationId:
                        type: string
                        example: eca1ba5ae1ca
                      name:
                        type: string
                        example: ILLUMINATION
                      description:
                        type: string
                        example: We curate and disseminate outstanding articles from diverse domains and
                          disciplines to create fusion and synergy.
                      slug:
                        type: string
                        example: illumination
                      creator:
                        type: string
                        example: dd3942a5498a
                      contributors:
                        type: array
                        items:
                          type: string
                          example: 3ed16e2a5be
                      tags:
                        type: array
                        items:
                          type: string
                          example: TECHNOLOGY
                      followersCount:
                        type: integer
                        format: int32
                        example: 69392
                      image:
                        type: string
                        example: https://miro.medium.com/undefined
                      logo:
                        type: string
                        example: https://miro.medium.com/1*AZxiin1Cvws3J0TwNUP2sQ.png
                      twitter:
                        type: string
                        example: digitalmehmet
                      facebook:
                        type: string
                        example: MediumIllumination
                      email:
                        type: string
                        example: myildizmel@gmail.com
  "/publications/{publicationId}/articles":
    get:
      summary: Get Publications's Articles
      tags:
        - Publications
      description: Returns the list of articles (`articleIds`) published in the publication.
      parameters:
        - name: publicationId
          in: path
          required: true
          description: Unique hash id assigned to every single Medium publication.
          schema:
            type: string
            example: eca1ba5ae1ca
        - name: count
          in: query
          required: false
          description: To limit the number of results.
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                      example: 97877675bbac
  "/publications/{publicationId}/newsletter":
    get:
      summary: Get Publication's Newsletter Info
      tags:
        - Publications
      description: >-
        
        Returns publication's newsletter related information like *Name, Description, Subscribers

        count, etc...*





        It takes `publicationId` as the path parameter.
      parameters:
        - name: publicationId
          in: path
          required: true
          description: It's a unique hash id assigned to every single Medium publication.
          schema:
            type: string
            example: eca1ba5ae1ca
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        example: ILLUMINATION Highlights
                      description:
                        type: string
                        example: We share our daily distribution list, featured stories, and progress
                          updates from our publication. Writer Inquires
                          https://digitalmehmet.com/contact We offer a free
                          Slack workspace for collaboration.
                      slug:
                        type: string
                        example: illumination-highlights
                      subscribersCount:
                        type: string
                        example: 4051
...